Improve model run error message clarity
This applies both for when an empty string or an incorrect model name
is specified. The prior had an error message already, but it was fairly
vague. The latter was using the default error output from later in
execution. This commit makes them consistent and more detailed,
providing two suggestions.