Prob don't need check-latest
"The action will first check the local cache for a version match. If a version is not found locally, it will pull it from the main branch of the go-versions repository. On miss or failure, it will fall back to downloading directly from go dist. To change the default behavior, please use the check-latest input." - https://github.com/actions/setup-go