Use models.github.ai as base url
* updates the base url
* updates relevant documentation
* updates display of list view
* updates how the models are displayed from a bare `run` or `view` command
* adds a FormatIdentifier func for consistently displaying model ID's in the places we do so

GitHub models is switching from using `https://models.inference.ai.azure.com/chat/completions` as the target
endpoint for inference. The suggested inference endpoint is now
`https://models.github.ai/inference/chat/completions`. `models.github.ai` expects the model param to be a composite
of the model's publisher and the model's name -- downcased and with `-` in pace of ` `.

The url change, and required param format change, made the changes to some verbiage and how we compare and display
model names (now called `ID` in the `list` output) a requirement as well.